chore: move useAllChainConfigsQuery from axelarscan to axelarConfigs
