+++ apps/maestro/src/pages/interchain-tokens/[tokenId]/index.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
 
 import { trpc } from "~/lib/trpc";
-import { useAllChainConfigsQuery } from "~/services/axelarscan/hooks";
+import { useAllChainConfigsQuery } from "~/services/axelarConfigs/hooks";
 import Page from "~/ui/layouts/Page";
 
 const TokenDetailsRedirectPage = () => {
