+++ apps/maestro/src/services/gmp/hooks.ts
@@ -2,8 +2,8 @@ import { Maybe } from "@axelarjs/utils";
 import { useMemo } from "react";
 
 import { trpc } from "~/lib/trpc";
-import { useAllChainConfigsQuery } from "../axelarscan/hooks";
 import { filterEligibleChains } from "~/lib/utils/chains";
+import { useAllChainConfigsQuery } from "~/services/axelarConfigs/hooks";
 
 export function useInterchainTokensQuery(input: {
   chainId?: number;
@@ -32,25 +32,34 @@ export function useInterchainTokensQuery(input: {
 
   const matchingTokens = data?.matchingTokens || [];
 
-  const destinationChainConfigs = matchingTokens.map((token) => {
-    return combinedComputed.indexedById[token.axelarChainId];
-  }).filter(Boolean);
+  const destinationChainConfigs = matchingTokens
+    .map((token) => {
+      return combinedComputed.indexedById[token.axelarChainId];
+    })
+    .filter(Boolean);
 
-  const eligibleChainConfigs = input.chainId ? filterEligibleChains(destinationChainConfigs, input.chainId) : destinationChainConfigs;
+  const eligibleChainConfigs = input.chainId
+    ? filterEligibleChains(destinationChainConfigs, input.chainId)
+    : destinationChainConfigs;
 
   return {
     ...queryResult,
     data: {
       ...data,
-      matchingTokens: data?.matchingTokens.filter((token) => {
-        return !!eligibleChainConfigs.find((x) => x.id === token.axelarChainId) || token.chainId === input.chainId;
-      }).map((token) => ({
-        ...token,
-        chain: combinedComputed.indexedById[token.axelarChainId ?? ""],
-        wagmiConfig: combinedComputed.wagmiChains?.find(
-          (x) => x?.id === Number(token.chainId)
-        ),
-      })),
+      matchingTokens: data?.matchingTokens
+        .filter((token) => {
+          return (
+            !!eligibleChainConfigs.find((x) => x.id === token.axelarChainId) ||
+            token.chainId === input.chainId
+          );
+        })
+        .map((token) => ({
+          ...token,
+          chain: combinedComputed.indexedById[token.axelarChainId ?? ""],
+          wagmiConfig: combinedComputed.wagmiChains?.find(
+            (x) => x?.id === Number(token.chainId)
+          ),
+        })),
       chain: Maybe.of(input.chainId).mapOrUndefined(
         (x) => combinedComputed.indexedByChainId[x]
       ),
