+++ apps/maestro/src/ui/layouts/Page.tsx
@@ -10,7 +10,7 @@ import { CHAIN_CONFIGS } from "~/config/chains";
 import RecentTransactions from "~/features/RecentTransactions/RecentTransactions";
 import SearchInterchainToken from "~/features/SearchInterchainToken";
 import { useAccount, useChainFromRoute, useSwitchChain } from "~/lib/hooks";
-import { useAllChainConfigsQuery } from "~/services/axelarscan/hooks";
+import { useAllChainConfigsQuery } from "~/services/axelarConfigs/hooks";
 import ChainsDropdown, { ChainIcon } from "~/ui/components/ChainsDropdown";
 import { ConditionalRenderInterchainBanner } from "../components/InterchainBanner";
 import ConnectWalletModal from "../compounds/ConnectWalletModal/ConnectWalletModal";
